Implemented user retrieval by random_slug (#66)
Resolves issue #https://github.com/cognoma/core-service/issues/58.
Inspired by PR https://github.com/cognoma/core-service/pull/65.

Note: while this was outside the scope of this commit, I did notice
that on the /users/ endpoint a list of all users is returned with no
authentication required. On the /classifiers/ endpoint a list of all
classifiers is returned with no authentication required. This seems
like a security concern.