Consolidate user endpoints, fix slug retrieval, update nginx CORS & config (#87)
* Separate id and slug user endpoints, fix slug retreival, update nginx CORS

When accessing a user, you can use either the slug or the id of the
user. The problem is that using a regex to capture a slug may make it
impossible for you to just use the id, as it may be interpreted as a
slug. The solution is to separate the two ways of access by prepending
the url with /slug/ or /id/ depending on the method used. For example,
accessing by slug would look like
`api.cognoma.org/users/slugs/alsihfakwrtakjh` whereas accessing by id
would look like `api.cognoma.org/users/id/8`.

Accessing by slug would also fail ungracefully if a user with the
specified slug didn’t exist. This commit handles that error gracefully
by returning a JSON 404 response with the message "No user exists with
random slug of {slug}."

Was facing some issues while working on the frontend with CORS on PUT
requests, so I updated the nginx dev config to fix these errors.

* Updated nginx configuration to retain request method checks

I was facing some issues with nginx not applying headers, and it seems
keeping in the conditional checks for the request method resolved those
issues.

* Removed users/id/ endpoint

Realized that there is no use in having two ways to access a user. The
only way that is needed is to access a user with the endpoint
/users/<random_slug>, because even if you were accessing the
/users/<id> endpoint from before you still would have needed to provide
the slug in the authorization header. Eliminating the /users/<id>
endpoint simplifies everything.